Add a more efficient builtin avrt_dummy_thread
The stack size is two above the minimum required to save and restore
state. It seems that the thread can be interrupted between calling
avrt_yield and disabling interrupts in the routine. Two extra bytes are
needed to store the return value of avrt_yield in addition to the saved
context. At least, that's what I think is happening.