Improve avrt_dummy_thread's stack usage
There is a tradeoff between two bytes of RAM and two bytes of flash. The
RAM is scarcer, so this change is an improvement. Unless, that is,
avrt_dummy_thread isn't used, which is admittedly likely. Meh. The
function only adds six bytes of flash. If it mattered, I could move it
to a separate compilation unit.